requirements.txt
pandas~=2.3.3
matplotlib~=3.10.7
pillow~=12.0.0
torchvision~=0.20.1
numpy~=2.2.6
scikit-learn~=1.7.2
torch~=2.5.1